Pass modifiers to arrow keys & enable quick braces
Active meta modifier was not getting passed to the arrow keys because the morph
layer was bound to meta to enable PgUp behaviour. Added them manually.

Typing braces on shift tap really fast was not possible because the closing
brace would activate the hold of the opening brace. Added a manual check for
this to insert the opening brace. One direction only to avoid juggling
modifier state to prevent infinite recursion.